formatting

+++ src/PriceConverter.sol
@@ -1,18 +1,18 @@
 // SPDX-License-Identifier: MIT
 pragma solidity ^0.8.18;
+
 import {AggregatorV3Interface} from "@chainlink/contracts/src/v0.8/shared/interfaces/AggregatorV3Interface.sol";
+
 library PriceConverter {
-function getPrice(AggregatorV3Interface priceFeed) internal view returns(uint256) {
+    function getPrice(AggregatorV3Interface priceFeed) internal view returns (uint256) {
         // AggregatorV3Interface dataFeed = AggregatorV3Interface(0x1b44F3514812d835EB1BDB0acB33d3fA3351Ee43);
-        (,int answer,,,) = priceFeed.latestRoundData();
-        return uint256(answer* 1e10);
+        (, int256 answer,,,) = priceFeed.latestRoundData();
+        return uint256(answer * 1e10);
     }
 
-    function convertedPrice(uint256 ethAmount, AggregatorV3Interface priceFeed) internal view returns(uint256) {
+    function convertedPrice(uint256 ethAmount, AggregatorV3Interface priceFeed) internal view returns (uint256) {
         uint256 ethPrice = getPrice(priceFeed);
         uint256 ethAmountUsd = (ethPrice * ethAmount) / 1e18;
         return ethAmountUsd;
-
     }
-
 }
